haiku_rag_slim/pyproject.toml
[project]

name = "haiku.rag-slim"
description = "Opinionated agentic RAG powered by LanceDB, Pydantic AI, and Docling - Minimal dependencies"
version = "0.20.0"
authors = [{ name = "Yiorgis Gozadinos", email = "ggozadinos@gmail.com" }]
license = { text = "MIT" }
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.12"
keywords = ["RAG", "lancedb", "vector-database", "ml", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]

dependencies = [
    "docling-core==2.54.0",
    "httpx>=0.28.1",
    "lancedb==0.25.3",
    "pathspec>=0.12.1",
    "pydantic>=2.12.5",
    "pydantic-ai-slim[openai,fastmcp,logfire,ag-ui]==1.27.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.19.2,<0.20.0",
    "watchfiles>=1.1.1",
]

[project.optional-dependencies]
# Document processing
docling = ["docling==2.64.0", "opencv-python-headless>=4.11.0.86"]
# Embedding providers
voyageai = ["voyageai>=0.3.5"]
# Rerankers
mxbai = ["mxbai-rerank>=0.1.6"]
cohere = ["cohere>=5.20.0"]
zeroentropy = ["zeroentropy>=0.1.0a7"]
# Inspector TUI
inspector = ["textual>=6.0.0", "textual-image>=0.8.4"]
# Model providers (delegated to pydantic-ai-slim)
anthropic = ["pydantic-ai-slim[anthropic]"]
groq = ["pydantic-ai-slim[groq]"]
google = ["pydantic-ai-slim[google]"]
mistral = ["pydantic-ai-slim[mistral]"]
bedrock = ["pydantic-ai-slim[bedrock]"]
vertexai = ["pydantic-ai-slim[vertexai]"]

[project.scripts]
haiku-rag = "haiku.rag.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["haiku"]
